reorganize abstract algebra module
