+++ src/qldpc/objects_test.py
@@ -139,7 +139,7 @@ def test_chain_complex(field: int = 3) -> None:
 
     # invalid chain complex constructions
     with pytest.raises(ValueError, match="inconsistent operator types"):
-        objects.ChainComplex([matrix, abstract.TrivialGroup.to_ring_array([[0]])])
+        objects.ChainComplex([matrix, abstract.RingArray.build([[0]])])
     with pytest.raises(ValueError, match="Inconsistent base fields"):
         objects.ChainComplex([galois.GF(field)(matrix)], field=field**2)
     with pytest.raises(ValueError, match="boundary operators .* must compose to zero"):
